[Refactor] Move shared files to common
