+++ tests/layers/jax/test_sharding.py
@@ -3,9 +3,9 @@ from unittest.mock import MagicMock
 
 import jax
 
-from tpu_inference.layers.jax.sharding import (Sharding, ShardingConfig,
-                                               ShardingRulesConfig,
-                                               ShardingStrategy)
+from tpu_inference.layers.common.sharding import (Sharding, ShardingConfig,
+                                                  ShardingRulesConfig,
+                                                  ShardingStrategy)
 
 
 class TestSharding(unittest.TestCase):
