+++ tpu_inference/layers/common/binary_search.py
